refactor(cactus-web): Move StatusLabel to its own component for re-use (as StatusMessage)
